third_party/blink/web_tests/fast/js/implicit-call-with-global-reentry-expected.txt

This test checks that implicit reentry to global code through a getter does not clobber the calling register file.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testObject.getterTest; testVar; is "PASS"
PASS testObject.setterTest = 1; testVar; is "PASS"
PASS testObject.toStringTest(); testVar; is "PASS"
PASS testObject.valueOfTest(); testVar; is "PASS"
PASS successfullyParsed is true

TEST COMPLETE
